+++ crates/rome_js_semantic/src/semantic_model/model.rs
@@ -1,5 +1,5 @@
 use super::*;
-use rome_js_syntax::{AnyJsRoot, JsFunctionDeclaration};
+use rome_js_syntax::{AnyJsFunction, AnyJsRoot, JsInitializerClause, JsVariableDeclarator};
 
 #[derive(Copy, Clone, Debug)]
 pub(crate) struct BindingIndex(usize);
@@ -336,11 +336,11 @@ impl SemanticModel {
         }
     }
 
-    /// Returns all [Call] of a [JsFunctionDeclaration].
+    /// Returns all [Call] of a [AnyJsFunction].
     ///
     /// ```rust
     /// use rome_rowan::{AstNode, SyntaxNodeCast};
-    /// use rome_js_syntax::{SourceType, JsFunctionDeclaration};
+    /// use rome_js_syntax::{SourceType, AnyJsFunction};
     /// use rome_js_semantic::{semantic_model, CallsExtensions, SemanticModelOptions};
     ///
     /// let r = rome_js_parser::parse("function f(){} f() f()", SourceType::js_module());
@@ -349,22 +349,31 @@ impl SemanticModel {
     /// let f_declaration = r
     ///     .syntax()
     ///     .descendants()
-    ///     .filter_map(JsFunctionDeclaration::cast)
+    ///     .filter_map(AnyJsFunction::cast)
     ///     .next()
     ///     .unwrap();
     ///
     /// let all_calls_to_f = model.all_calls(&f_declaration);
-    /// assert_eq!(2, all_calls_to_f.count());
+    /// assert_eq!(2, all_calls_to_f.unwrap().count());
     /// // or
     /// let all_calls_to_f = f_declaration.all_calls(&model);
-    /// assert_eq!(2, all_calls_to_f.count());
+    /// assert_eq!(2, all_calls_to_f.unwrap().count());
     /// ```
-    pub fn all_calls(&self, function: &JsFunctionDeclaration) -> AllCallsIter {
-        let references = function
-            .id()
-            .map(|id| id.all_reads(self))
-            .unwrap_or_else(|_| std::iter::successors(None, Reference::find_next_read));
+    pub fn all_calls(&self, function: &AnyJsFunction) -> Option<AllCallsIter> {
+        let identifier = match function {
+            AnyJsFunction::JsFunctionDeclaration(declaration) => declaration.id().ok()?,
+            AnyJsFunction::JsFunctionExportDefaultDeclaration(declaration) => declaration.id()?,
+            AnyJsFunction::JsArrowFunctionExpression(_)
+            | AnyJsFunction::JsFunctionExpression(_) => {
+                let parent = function
+                    .parent::<JsInitializerClause>()?
+                    .parent::<JsVariableDeclarator>()?;
+                parent.id().ok()?.as_any_js_binding()?.clone()
+            }
+        };
 
-        AllCallsIter { references }
+        Some(AllCallsIter {
+            references: identifier.all_reads(self),
+        })
     }
 }